Fix #428 - speed up package builds
Roughly halve the build time (down to about 140 seconds on my machine)

Python is now only built when either the @deps/python@ directory doesn't exist or you passed the @--with-python@ switch to @rhel-build.sh@. Instead of compiling Python inside @rpmbuild@ as we did before, we simply unpack a bzipped tar archive containing the binaries which are built in the build script. This enables us to speed up builds without kludging along with the system's Python interpreter.